% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{clone.foo::instance}
\alias{clone.foo::instance}
\title{Clone an Object Instance}
\usage{
\method{clone}{foo::instance}(obj, deep = TRUE)
}
\arguments{
\item{deep}{}
}
\description{
Clone an Object Instance
}
