% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clone.Q7instance}
\alias{clone.Q7instance}
\title{Clone an Instance}
\usage{
\method{clone}{Q7instance}(inst, deep = TRUE, ...)
}
\arguments{
\item{inst}{Q7 object instance}

\item{deep}{to copy nested object instances recursively; Boolean}

\item{...}{dot-dot-dot}
}
\value{
Q7 object instance
}
\description{
Clone an Instance
}
\examples{

Type1 <- type(function(num){
  print_num <- function(){
    base::print(num)
  }
})
myType1 <- Type1(1)
myType1$print_num()
myType1_clone <- clone(myType1)
myType1_clone$print_num()

}
