% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{feature_generic}
\alias{feature_generic}
\title{Create a Generic Feature}
\usage{
feature_generic(s3, ...)
}
\arguments{
\item{s3}{S3 Class of the feature}

\item{...}{dot-dot-dot}
}
\value{
a generic Q7 feature
}
\description{
Use this function when you need to create more than one methods
for Q7 types with different S3 classes.
The \code{s3} field and the feature's name should be the same.
}
\seealso{
\code{\link{feature}}
}
