% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is}
\alias{is}
\alias{is_type}
\alias{is_instance}
\alias{is_feature}
\title{Is it a Q7 Type, Instance or Feature?}
\usage{
is_type(x)

is_instance(x)

is_feature(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean
}
\description{
Is it a Q7 Type, Instance or Feature?
}
