% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{list2inst}
\alias{list2inst}
\title{Build an foo::instance from a list}
\usage{
list2inst(x, s3_class = "default", parent = parent.frame(), ...)
}
\arguments{
\item{...}{}
}
\description{
Build an foo::instance from a list
}
