% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{localize}
\alias{localize}
\title{Make a Localized Copy of a Q7 Type or Instance}
\usage{
localize(obj, envir = parent.frame())
}
\arguments{
\item{obj}{Q7 type or instance}

\item{envir}{environment}
}
\value{
function
}
\description{
Make a Localized Copy of a Q7 Type or Instance
}
