% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{type}
\alias{type}
\title{Create an foo Type}
\usage{
type(fn = function() { }, s3_class = "default")
}
\arguments{
\item{fn}{a constructor function}
}
\value{
foo::type, function
}
\description{
This function is a object generator
}
