% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{localize}
\alias{localize}
\title{Localize a Function's Environment}
\usage{
localize(fn, envir = parent.frame())
}
\arguments{
\item{fn}{function}

\item{envir}{environment}
}
\value{
function
}
\description{
Localize a Function's Environment
}
