% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{localize}
\alias{localize}
\title{Localize a Type's Environment}
\usage{
localize(type, envir = parent.frame())
}
\arguments{
\item{type}{foo::type}

\item{envir}{environment}
}
\value{
foo::type
}
\description{
Localize a Type's Environment
}
