% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/utils.R
\name{type}
\alias{type}
\alias{print.Q7type}
\alias{print.Q7instance}
\alias{is_type}
\alias{is_instance}
\title{Create an Q7 Type}
\usage{
type(fn = function() { }, s3 = "default")

\method{print}{Q7type}(x, ...)

\method{print}{Q7instance}(x, ...)

is_type(x)

is_instance(x)
}
\arguments{
\item{fn}{a constructor function}

\item{s3}{S3 class of the object}
}
\value{
Q7type, function
}
\description{
This function is a object generator
}
\examples{

}
