% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{type}
\alias{type}
\title{Create an Q7 Type}
\usage{
type(fn = function() { }, s3 = "default")
}
\arguments{
\item{fn}{a constructor function}
}
\value{
Q7::type, function
}
\description{
This function is a object generator
}
\examples{

}
