% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{type}
\alias{type}
\title{Create a Q7 Type}
\usage{
type(x = function() {
}, s3 = "Q7default")
}
\arguments{
\item{x}{function or expression; becomes the definition of the obcject}

\item{s3}{S3 class of the object; necessary when using}
}
\value{
Q7 type; function
}
\description{
Create a Q7 Type
}
\examples{

Adder <- type(function(num1, num2){
    add_nums <- function(){
        num1 + num2
    }
 })

myAdder <- Adder(1, 2)
myAdder$add_nums()

}
